v0.1.0: download works!
There are a few things that needs to be modified on original API:
a) Name: we need to know the name .plg file needs to be
b) Download url: must be a .plg DIRECT download link. If not, it doesn't
download.
c) If url is a direct download link ending with xxxx.plg or
xxxx.plg?raw=true we can get a) from here instead. Also, I recommend to
move all plugins to our host repo

WIP: Use default layout with progress bar and parse "game" titleID