installtitledb.c: re-draw all original UI with download method
It have still some bugs, but because of I'm a bit bussy in this moment,
I'll upload to github to have it as a backup. This is a WIP and I'm
still working on it!

Additions:
-When user press Install, a dialog ask if user want to download the
plugin. Click NO to return to title list. When pressing A, plugin will
download. NTRDBI will notify when download finished (exceptions aren't
handle right now because this is a WIP and not suitable for release yet.

Bugs (they aren't actually)
- Okay (Any button) -> keeps downloading.
Actually this isn't a bug. It's because I haven't handle it either.

I'm still working on this file and there are a lot of things more to
implement here!